Move adapter classes to "adapters" package
 * Make Wallet comparable and move in to "entities"
 * Cleanup code
