--- app/src/main/java/io/horizontalsystems/bankwallet/core/Wallet.kt
@@ -1,7 +1,0 @@
-package io.horizontalsystems.bankwallet.core
-
-import io.horizontalsystems.bankwallet.entities.Account
-import io.horizontalsystems.bankwallet.entities.Coin
-import io.horizontalsystems.bankwallet.entities.SyncMode
-
-class Wallet(val coin: Coin, val account: Account, val syncMode: SyncMode?)
