+++ app/src/main/java/io/horizontalsystems/bankwallet/core/managers/DefaultWalletCreator.kt
@@ -3,9 +3,9 @@ package io.horizontalsystems.bankwallet.core.managers
 import io.horizontalsystems.bankwallet.core.IAppConfigProvider
 import io.horizontalsystems.bankwallet.core.IWalletFactory
 import io.horizontalsystems.bankwallet.core.IWalletManager
-import io.horizontalsystems.bankwallet.core.Wallet
 import io.horizontalsystems.bankwallet.entities.Account
 import io.horizontalsystems.bankwallet.entities.AccountType
+import io.horizontalsystems.bankwallet.entities.Wallet
 import io.horizontalsystems.bankwallet.modules.transactions.CoinCode
 
 class DefaultWalletCreator(private val walletManager: IWalletManager, private val appConfigProvider: IAppConfigProvider, private val walletFactory: IWalletFactory) {
