+++ app/src/main/java/io/horizontalsystems/bankwallet/core/managers/EosKitManager.kt
@@ -1,7 +1,11 @@
 package io.horizontalsystems.bankwallet.core.managers
 
-import io.horizontalsystems.bankwallet.core.*
+import io.horizontalsystems.bankwallet.core.App
+import io.horizontalsystems.bankwallet.core.IAppConfigProvider
+import io.horizontalsystems.bankwallet.core.IEosKitManager
+import io.horizontalsystems.bankwallet.core.UnsupportedAccountException
 import io.horizontalsystems.bankwallet.entities.AccountType
+import io.horizontalsystems.bankwallet.entities.Wallet
 import io.horizontalsystems.eoskit.EosKit
 
 class EosKitManager(appConfig: IAppConfigProvider) : IEosKitManager {
