+++ app/src/main/java/io/horizontalsystems/bankwallet/core/managers/WalletStorage.kt
@@ -3,6 +3,7 @@ package io.horizontalsystems.bankwallet.core.managers
 import io.horizontalsystems.bankwallet.core.*
 import io.horizontalsystems.bankwallet.entities.Account
 import io.horizontalsystems.bankwallet.entities.EnabledWallet
+import io.horizontalsystems.bankwallet.entities.Wallet
 
 class WalletStorage(
         private val appConfigProvider: IAppConfigProvider,
