+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/managecoins/ManageWalletsInteractor.kt
@@ -4,6 +4,7 @@ import io.horizontalsystems.bankwallet.core.*
 import io.horizontalsystems.bankwallet.entities.AccountType
 import io.horizontalsystems.bankwallet.entities.Coin
 import io.horizontalsystems.bankwallet.entities.SyncMode
+import io.horizontalsystems.bankwallet.entities.Wallet
 
 class ManageWalletsInteractor(
         private val appConfigProvider: IAppConfigProvider,
