+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/managecoins/ManageWalletsPresenter.kt
@@ -1,10 +1,10 @@
 package io.horizontalsystems.bankwallet.modules.managecoins
 
 import io.horizontalsystems.bankwallet.core.DefaultAccountType
-import io.horizontalsystems.bankwallet.core.Wallet
 import io.horizontalsystems.bankwallet.entities.AccountType
 import io.horizontalsystems.bankwallet.entities.Coin
 import io.horizontalsystems.bankwallet.entities.SyncMode
+import io.horizontalsystems.bankwallet.entities.Wallet
 
 class ManageWalletsPresenter(private val interactor: ManageWalletsModule.IInteractor, private val router: ManageWalletsModule.IRouter)
     : ManageWalletsModule.IViewDelegate, ManageWalletsModule.IInteractorDelegate {
