+++ app/src/test/java/io/horizontalsystems/bankwallet/core/managers/RateSyncerTest.kt
@@ -4,9 +4,9 @@ import com.nhaarman.mockito_kotlin.*
 import io.horizontalsystems.bankwallet.core.IAdapter
 import io.horizontalsystems.bankwallet.core.IAdapterManager
 import io.horizontalsystems.bankwallet.core.ICurrencyManager
-import io.horizontalsystems.bankwallet.core.Wallet
 import io.horizontalsystems.bankwallet.entities.Coin
 import io.horizontalsystems.bankwallet.entities.Currency
+import io.horizontalsystems.bankwallet.entities.Wallet
 import io.horizontalsystems.bankwallet.modules.RxBaseTest
 import io.reactivex.subjects.BehaviorSubject
 import io.reactivex.subjects.PublishSubject
